Small corrections as well as changes for TeXLive20
TeXLive2020 doesn't like luapackageloader anymore (at least not the way
I used it), so out of necessity I found a way to let it work with
kpathsea (put the module folder in texmf/scripts/kpsewhich/). However it
doesn't like it if the main file is called init.lua instead of
MODULENAME.lua, so I had to change that.