Clean up handling of dotless i
Previous commit unnecessarily added a second round replacement of i to
dotless i, even though that is already handled by the default conversion
functions. Of the current schemes only Japanese ones need to handle them
themselves because they override the default conversion function for a
slight boost in performance. (So do Korean ones but none of the Korean
schemes which are currently supported features i as a carrier for
diacritics.)