Bug and documentation fixes
Fixed a bug which caused any replacement to be performed only once per
unit (i.e. per slice for languages that perform separations; otherwise
only once in the whole string). Also fixed a few minor mistakes in the
documentation that happened during the last commit and updated the
license to include added files